+++ babel.config.js
@@ -1,11 +1,5 @@
 module.exports = {
   presets: [
     '@babel/preset-env',
-    '@babel/preset-typescript',
-  ],
-  plugins: [
-    '@babel/plugin-proposal-object-rest-spread',
-    '@babel/plugin-syntax-dynamic-import',
-    '@babel/plugin-proposal-class-properties',
   ],
 };